+++ coreml_diffusion/conversion/attention.py
