+++ coreml_diffusion/conversion/trace.py
