+++ coreml_diffusion/logger.py
@@ -1,0 +1,5 @@
+import logging
+
+logging.basicConfig()
+logger = logging.getLogger(__name__)
+logger.setLevel(logging.INFO)
